requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
PyJWT==2.9.0
python-dotenv==1.0.1
pyzbar==0.1.8
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
Werkzeug==3.1.3
wheel==0.45.1
